Move test result parsing to generators (#105)
* temp change

* refactor: move test result parsing to generators

* refactor: fix formatting problems
